PP-2890 Use default equals and hashcode instead of guava's
