+++ src/main/java/uk/gov/pay/directdebit/mandate/resources/ConfirmPaymentResource.java
@@ -3,7 +3,6 @@ package uk.gov.pay.directdebit.mandate.resources;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.gov.pay.directdebit.gatewayaccounts.model.GatewayAccount;
-import uk.gov.pay.directdebit.mandate.services.PaymentConfirmService;
 import uk.gov.pay.directdebit.payments.model.DirectDebitPaymentProvider;
 import uk.gov.pay.directdebit.payments.model.PaymentProviderFactory;
